+++ src/main.js
@@ -1,0 +1,3 @@
+import vcSelect from './Select'
+
+module.exports = vcSelect